+++ puffin/src/lib.rs
@@ -98,10 +98,12 @@ pub use data::*;
 pub use merge::*;
 pub use profile_view::{select_slowest, FrameView, GlobalFrameView};
 
+use parking_lot::{Mutex, RwLock};
 use std::collections::BTreeMap;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
-use std::sync::Mutex;
+use std::sync::{
+    atomic::{AtomicBool, Ordering},
+    Arc,
+};
 
 static MACROS_ON: AtomicBool = AtomicBool::new(false);
 
@@ -321,23 +323,32 @@ pub struct StreamInfoRef<'a> {
 
 // ----------------------------------------------------------------------------
 
-/// One frame worth of profile data, collected from many sources.
-#[derive(Clone)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub struct FrameData {
-    pub frame_index: FrameIndex,
-    pub thread_streams: BTreeMap<ThreadInfo, Arc<StreamInfo>>,
-    pub range_ns: (NanoSecond, NanoSecond),
-    /// The uncompressed size
-    pub num_bytes: usize,
-    pub num_scopes: usize,
+pub type ThreadStreams = BTreeMap<ThreadInfo, Arc<StreamInfo>>;
 
-    /// Size of the compressed data (in bytes), if this was decompressed.
-    #[cfg_attr(feature = "serde", serde(default, skip))]
-    pub compressed_size: Option<usize>,
+/// Meta-information about a frame.
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+#[derive(Clone, Debug)]
+pub struct FrameMeta {
+    /// What frame this is (counting from 0 at application startup).
+    pub frame_index: FrameIndex,
+    /// The span we cover.
+    pub range_ns: (NanoSecond, NanoSecond),
+    /// The unpacked size of all streams.
+    pub num_bytes: usize,
+    /// Total number of scopes.
+    pub num_scopes: usize,
 }
 
-impl FrameData {
+/// One frame worth of profile data, collected from many sources.
+///
+/// More often encoded as a [`FrameData`].
+pub struct UnpackedFrameData {
+    pub meta: FrameMeta,
+    /// `None` if still compressed.
+    pub thread_streams: ThreadStreams,
+}
+
+impl UnpackedFrameData {
     pub fn new(
         frame_index: FrameIndex,
         thread_streams: BTreeMap<ThreadInfo, StreamInfo>,
@@ -361,42 +372,205 @@ impl FrameData {
 
         if min_ns <= max_ns {
             Ok(Self {
-                frame_index,
+                meta: FrameMeta {
+                    frame_index,
+                    range_ns: (min_ns, max_ns),
+                    num_bytes,
+                    num_scopes,
+                },
                 thread_streams,
-                range_ns: (min_ns, max_ns),
-                num_bytes,
-                num_scopes,
-                compressed_size: None,
             })
         } else {
             Err(Error::Empty)
         }
     }
 
+    pub fn frame_index(&self) -> u64 {
+        self.meta.frame_index
+    }
+
+    pub fn range_ns(&self) -> (NanoSecond, NanoSecond) {
+        self.meta.range_ns
+    }
+
     pub fn duration_ns(&self) -> NanoSecond {
-        let (min, max) = self.range_ns;
+        let (min, max) = self.meta.range_ns;
+        max - min
+    }
+}
+
+/// One frame worth of profile data, collected from many sources.
+///
+/// This has interior mutability with double storage:
+/// * Unpacked data ([`UnpackedFrameData`])
+/// * Packed (compressed) data
+///
+/// One or both are always stored.
+/// This allows RAM-efficient storage and viewing of many frames of profiling data.
+/// Packing and unpacking is done lazily, on-demand.
+pub struct FrameData {
+    pub meta: FrameMeta,
+    /// * `None` if still compressed.
+    /// * `Some(Err(…))` if there was a problem during unpacking.
+    /// * `Some(Ok(…))` if unpacked.
+    unpacked_frame: RwLock<Option<anyhow::Result<Arc<UnpackedFrameData>>>>,
+    /// [`UnpackedFrameData::thread_streams`], compressed with zstd.
+    /// `None` if not yet compressed.
+    packed_zstd_streams: RwLock<Option<Vec<u8>>>,
+}
+
+impl FrameData {
+    pub fn new(
+        frame_index: FrameIndex,
+        thread_streams: BTreeMap<ThreadInfo, StreamInfo>,
+    ) -> Result<Self> {
+        Ok(Self::from_unpacked(Arc::new(UnpackedFrameData::new(
+            frame_index,
+            thread_streams,
+        )?)))
+    }
+
+    pub fn from_unpacked(frame: Arc<UnpackedFrameData>) -> Self {
+        Self {
+            meta: frame.meta.clone(),
+            unpacked_frame: RwLock::new(Some(Ok(frame))),
+            packed_zstd_streams: RwLock::new(None),
+        }
+    }
+
+    pub fn frame_index(&self) -> u64 {
+        self.meta.frame_index
+    }
+
+    pub fn range_ns(&self) -> (NanoSecond, NanoSecond) {
+        self.meta.range_ns
+    }
+
+    pub fn duration_ns(&self) -> NanoSecond {
+        let (min, max) = self.meta.range_ns;
         max - min
     }
 
-    /// Writes one Frame into a stream, prefixed by it's length (u32 le).
+    /// Number of bytes used by the compressed data, if compressed.
+    pub fn packed_size(&self) -> Option<usize> {
+        self.packed_zstd_streams.read().as_ref().map(|c| c.len())
+    }
+
+    /// Number of bytes used when unpacked, if known.
+    pub fn unpacked_size(&self) -> Option<usize> {
+        if self.has_unpacked() {
+            Some(self.meta.num_bytes)
+        } else {
+            None
+        }
+    }
+
+    /// bytes currently used by the unpacked and compressed data.
+    pub fn bytes_of_ram_used(&self) -> usize {
+        self.unpacked_size().unwrap_or(0) + self.packed_size().unwrap_or(0)
+    }
+
+    /// Do we have a packed (compressed) version stored internally?
+    pub fn has_packed(&self) -> bool {
+        self.packed_zstd_streams.read().is_some()
+    }
+
+    /// Do we have a unpacked version stored internally?
+    pub fn has_unpacked(&self) -> bool {
+        self.unpacked_frame.read().is_some()
+    }
+
+    /// Lazily unpacks.
+    pub fn unpack(&self) -> anyhow::Result<Arc<UnpackedFrameData>> {
+        fn unpack_frame_data(
+            meta: FrameMeta,
+            compressed: &[u8],
+        ) -> anyhow::Result<UnpackedFrameData> {
+            use anyhow::Context as _;
+            use bincode::Options as _;
+
+            let streams_serialized = zstd::decode_all(compressed).context("zstd decompress")?;
+
+            let thread_streams: ThreadStreams = bincode::options()
+                .deserialize(&streams_serialized)
+                .context("bincode deserialize")?;
+
+            Ok(UnpackedFrameData {
+                meta,
+                thread_streams,
+            })
+        }
+
+        let needs_unpack = self.unpacked_frame.read().is_none();
+        if needs_unpack {
+            let packed_lock = self.packed_zstd_streams.read();
+            let packed = packed_lock
+                .as_ref()
+                .expect("FrameData is neither packed or unpacked");
+
+            let frame_data_result = unpack_frame_data(self.meta.clone(), packed);
+            let frame_data_result = frame_data_result.map(Arc::new);
+            *self.unpacked_frame.write() = Some(frame_data_result);
+        }
+
+        match self.unpacked_frame.read().as_ref().unwrap() {
+            Ok(frame) => Ok(frame.clone()),
+            Err(err) => Err(anyhow::format_err!("{}", err)), // can't clone `anyhow::Error`
+        }
+    }
+
+    /// Make the [`FrameData`] use up less memory.
+    /// Idempotent.
+    pub fn pack(&self) {
+        self.created_packed();
+        *self.unpacked_frame.write() = None;
+    }
+
+    /// Create a packed storage without freeing the unpacked storage.
+    fn created_packed(&self) {
+        use bincode::Options as _;
+        let needs_compress = self.packed_zstd_streams.read().is_none();
+        if needs_compress {
+            let unpacked_frame = self
+                .unpacked_frame
+                .read()
+                .as_ref()
+                .expect("We should have an unpacked frame if we don't have a packed one")
+                .as_ref()
+                .expect("The unpacked frame should be error free, since it doesn't come from packed source")
+                .clone();
+
+            let streams_serialized = bincode::options()
+                .serialize(&unpacked_frame.thread_streams)
+                .expect("bincode failed to encode");
+
+            // zstd cuts sizes in half compared to lz4_flex
+            let level = 3;
+            let streams_compressed =
+                zstd::encode_all(std::io::Cursor::new(&streams_serialized), level)
+                    .expect("zstd failed to compress");
+
+            *self.packed_zstd_streams.write() = Some(streams_compressed);
+        }
+    }
+
+    /// Writes one [`FrameData`] into a stream, prefixed by it's length (u32 le).
     #[cfg(feature = "serialization")]
     pub fn write_into(&self, write: &mut impl std::io::Write) -> anyhow::Result<()> {
         use bincode::Options as _;
-        let serialized = bincode::options().serialize(self)?;
+        let meta_serialized = bincode::options().serialize(&self.meta)?;
 
-        if false {
-            let compressed = lz4_flex::compress_prepend_size(&serialized);
-            write.write_all(b"PFD0")?;
-            write.write_all(&(compressed.len() as u32).to_le_bytes())?;
-            write.write_all(&compressed)?;
-        } else {
-            // zstd cuts sizes in half compared to lz4_flex
-            let level = 3;
-            let compressed = zstd::encode_all(std::io::Cursor::new(&serialized), level)?;
-            write.write_all(b"PFD1")?;
-            write.write_all(&(compressed.len() as u32).to_le_bytes())?;
-            write.write_all(&compressed)?;
-        }
+        write.write_all(b"PFD2")?;
+        write.write_all(&(meta_serialized.len() as u32).to_le_bytes())?;
+        write.write_all(&meta_serialized)?;
+
+        self.pack();
+        let zstd_streams_lock = self.packed_zstd_streams.read();
+        let zstd_streams = zstd_streams_lock.as_ref().unwrap();
+
+        write.write_all(&(zstd_streams.len() as u32).to_le_bytes())?;
+        write.write_all(zstd_streams)?;
+
         Ok(())
     }
 
@@ -407,6 +581,7 @@ impl FrameData {
     #[cfg(feature = "serialization")]
     pub fn read_next(read: &mut impl std::io::Read) -> anyhow::Result<Option<Self>> {
         use anyhow::Context as _;
+        use bincode::Options as _;
 
         let mut header = [0_u8; 4];
         if let Err(err) = read.read_exact(&mut header) {
@@ -414,6 +589,41 @@ impl FrameData {
                 return Ok(None);
             } else {
                 return Err(err.into());
+            }
+        }
+
+        #[derive(Clone)]
+        #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+        pub struct LegacyFrameData {
+            pub frame_index: FrameIndex,
+            pub thread_streams: ThreadStreams,
+            pub range_ns: (NanoSecond, NanoSecond),
+            pub num_bytes: usize,
+            pub num_scopes: usize,
+        }
+
+        impl LegacyFrameData {
+            fn into_unpacked_frame_data(self) -> UnpackedFrameData {
+                let Self {
+                    frame_index,
+                    thread_streams,
+                    range_ns,
+                    num_bytes,
+                    num_scopes,
+                } = self;
+                UnpackedFrameData {
+                    meta: FrameMeta {
+                        frame_index,
+                        range_ns,
+                        num_bytes,
+                        num_scopes,
+                    },
+                    thread_streams,
+                }
+            }
+
+            fn into_frame_data(self) -> FrameData {
+                FrameData::from_unpacked(Arc::new(self.into_unpacked_frame_data()))
             }
         }
 
@@ -430,13 +640,12 @@ impl FrameData {
                 let serialized =
                     lz4_flex::decompress_size_prepended(&compressed).context("lz4 decompress")?;
 
-                use bincode::Options as _;
-                let mut frame: Self = bincode::options()
+                let legacy: LegacyFrameData = bincode::options()
                     .deserialize(&serialized)
                     .context("bincode deserialize")?;
-                frame.compressed_size = Some(compressed_length);
-                Ok(Some(frame))
+                Ok(Some(legacy.into_frame_data()))
             } else if &header == b"PFD1" {
+                // Added 2021-09
                 let mut compressed_length = [0_u8; 4];
                 read.read_exact(&mut compressed_length)?;
                 let compressed_length = u32::from_le_bytes(compressed_length) as usize;
@@ -445,26 +654,49 @@ impl FrameData {
 
                 let serialized = zstd::decode_all(&compressed[..]).context("zstd decompress")?;
 
-                use bincode::Options as _;
-                let mut frame: Self = bincode::options()
+                let legacy: LegacyFrameData = bincode::options()
                     .deserialize(&serialized)
                     .context("bincode deserialize")?;
-                frame.compressed_size = Some(compressed_length);
-                Ok(Some(frame))
+                Ok(Some(legacy.into_frame_data()))
+            } else if &header == b"PFD2" {
+                // Added 2021-11-15
+                let mut meta_length = [0_u8; 4];
+                read.read_exact(&mut meta_length)?;
+                let meta_length = u32::from_le_bytes(meta_length) as usize;
+                let mut meta = vec![0_u8; meta_length];
+                read.read_exact(&mut meta)?;
+
+                let meta: FrameMeta = bincode::options()
+                    .deserialize(&meta)
+                    .context("bincode deserialize")?;
+
+                let mut streams_compressed_length = [0_u8; 4];
+                read.read_exact(&mut streams_compressed_length)?;
+                let streams_compressed_length =
+                    u32::from_le_bytes(streams_compressed_length) as usize;
+                let mut streams_compressed = vec![0_u8; streams_compressed_length];
+                read.read_exact(&mut streams_compressed)?;
+
+                // Don't unpack now - do it if/when needed!
+
+                Ok(Some(Self {
+                    meta,
+                    unpacked_frame: RwLock::new(None),
+                    packed_zstd_streams: RwLock::new(Some(streams_compressed)),
+                }))
             } else {
                 anyhow::bail!("Failed to decode: this data is newer than this reader. Please update your puffin version!");
             }
         } else {
-            // Old packet without magic header
+            // Very old packet without magic header
             let mut bytes = vec![0_u8; u32::from_le_bytes(header) as usize];
             read.read_exact(&mut bytes)?;
 
             use bincode::Options as _;
-            Ok(Some(
-                bincode::options()
-                    .deserialize(&bytes)
-                    .context("bincode deserialize")?,
-            ))
+            let legacy: LegacyFrameData = bincode::options()
+                .deserialize(&bytes)
+                .context("bincode deserialize")?;
+            Ok(Some(legacy.into_frame_data()))
         }
     }
 }
@@ -598,10 +830,10 @@ impl Default for GlobalProfiler {
 
 impl GlobalProfiler {
     /// Access to the global profiler singleton.
-    pub fn lock() -> std::sync::MutexGuard<'static, Self> {
+    pub fn lock() -> parking_lot::MutexGuard<'static, Self> {
         use once_cell::sync::Lazy;
         static GLOBAL_PROFILER: Lazy<Mutex<GlobalProfiler>> = Lazy::new(Default::default);
-        GLOBAL_PROFILER.lock().unwrap() // panic on mutex poisoning
+        GLOBAL_PROFILER.lock()
     }
 
     /// You need to call this once at the start of every frame.